+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 TBD: Version 1.1 (??? 2018)
 --------------------------------------
+New Rules
+   - #247: MissingOverrideAnnotation rule (enhanced) - Checks for methods that override a method in a super class or implement a method in an interface but are not annotated with @Override. (Marcin Erdmann)
+
 Updated/Enhanced Rules and Bug Fixes
    - #239: Ignore super calls in UnnecessarySetterRule. (Marcin Erdmann) 
    - #238: Update Gradle to 4.2. (Marcin Erdmann)